fix: added forward ref to the last components
