+++ src/components/alert/Alert.tsx
@@ -1,91 +1,85 @@
 'use client';
-import React, { FC, SVGProps, useState, useEffect } from 'react';
 import { Alert as FlowbiteAlert } from 'flowbite-react';
+import React, { forwardRef, useEffect, useState } from 'react';
 import { HiInformationCircle } from 'react-icons/hi';
 
-export type iAlertProps = {
-  id?: string;
-  additionalContent?: JSX.Element;
-  role?: string;
-  icon?: FC<SVGProps<SVGSVGElement>>;
-  className?: string;
-  children: JSX.Element | string;
-  testId?: string;
-  onClick?: () => void;
-  onDismiss?: () => void;
-  onMouseEnter?: () => void;
-  onMouseLeave?: () => void;
-  dismissAfter?: number; // Optional time in milliseconds to automatically dismiss the alert
-};
+import { iAlertProps } from '@/types/types/components/components';
 
-const Alert = ({
-  id,
-  additionalContent,
-  role = 'blue',
-  icon = HiInformationCircle,
-  className,
-  children,
-  testId,
-  onClick,
-  onDismiss,
-  onMouseEnter,
-  onMouseLeave,
-  dismissAfter,
-  ...rest
-}: iAlertProps) => {
-  const [showAlert, setShowAlert] = useState(true);
-  const [isMounted, setIsMounted] = useState(true);
-  const transitionDurations = 300;
-  useEffect(() => {
-    let timeoutId: NodeJS.Timeout;
+const Alert = forwardRef<any, iAlertProps>(
+  (
+    {
+      id,
+      additionalContent,
+      role = 'blue',
+      icon = HiInformationCircle,
+      className,
+      children,
+      testId,
+      onClick,
+      onDismiss,
+      onMouseEnter,
+      onMouseLeave,
+      dismissAfter,
+      ...rest
+    }: iAlertProps,
+    ref: React.Ref<any> | null,
+  ) => {
+    const [showAlert, setShowAlert] = useState(true);
+    const [isMounted, setIsMounted] = useState(true);
+    const transitionDurations = 300;
 
-    if (dismissAfter) {
-      timeoutId = setTimeout(() => {
-        setShowAlert(false);
-        if (onDismiss) onDismiss();
-        setTimeout(() => {
-          setIsMounted(false);
-        }, transitionDurations);
-      }, dismissAfter);
-    }
+    useEffect(() => {
+      let timeoutId: NodeJS.Timeout;
 
-    return () => {
-      if (timeoutId) clearTimeout(timeoutId);
+      if (dismissAfter) {
+        timeoutId = setTimeout(() => {
+          setShowAlert(false);
+          if (onDismiss) onDismiss();
+          setTimeout(() => {
+            setIsMounted(false);
+          }, transitionDurations);
+        }, dismissAfter);
+      }
+
+      return () => {
+        if (timeoutId) clearTimeout(timeoutId);
+      };
+    }, [dismissAfter, id, onDismiss]);
+
+    const handleDismiss = () => {
+      setShowAlert(false);
+      if (onDismiss) onDismiss();
+      setTimeout(() => {
+        setIsMounted(false);
+      }, transitionDurations);
     };
-  }, [dismissAfter, id, onDismiss]);
 
-  const handleDismiss = () => {
-    setShowAlert(false);
-    if (onDismiss) onDismiss();
-    setTimeout(() => {
-      setIsMounted(false);
-    }, transitionDurations);
-  };
-
-  return (
-    isMounted && (
-      <div
-        className={`transition-all duration-${transitionDurations} ease ${showAlert ? 'max-h-40 opacity-1' : 'max-h-0 opacity-0'}`}
-      >
-        <FlowbiteAlert
-          id={id}
-          className={className}
-          additionalContent={additionalContent}
-          color={role}
-          icon={icon}
-          rounded
-          data-testid={testId}
-          onClick={onClick}
-          onDismiss={handleDismiss}
-          onMouseEnter={onMouseEnter}
-          onMouseLeave={onMouseLeave}
-          {...rest}
+    return (
+      isMounted && (
+        <div
+          className={`transition-all duration-${transitionDurations} ease ${showAlert ? 'max-h-40 opacity-1' : 'max-h-0 opacity-0'}`}
         >
-          {children}
-        </FlowbiteAlert>
-      </div>
-    )
-  );
-};
+          <FlowbiteAlert
+            id={id}
+            className={className}
+            additionalContent={additionalContent}
+            color={role}
+            icon={icon}
+            rounded
+            data-testid={testId}
+            onClick={onClick}
+            onDismiss={handleDismiss}
+            onMouseEnter={onMouseEnter}
+            onMouseLeave={onMouseLeave}
+            ref={ref}
+            {...rest}
+          >
+            {children}
+          </FlowbiteAlert>
+        </div>
+      )
+    );
+  },
+);
 
 export default Alert;
