+++ src/components/buttonGroup/ButtonGroup.tsx
@@ -1,42 +1,39 @@
-import React, { ReactNode } from 'react';
 import { Button as FlowbiteButton } from 'flowbite-react';
+import React, { forwardRef } from 'react';
 
-export type iButtonGroupProps = {
-  id?: string;
-  outline?: boolean;
-  className?: string;
-  children: React.ReactNode;
-  testId?: string;
-  onClick?: () => void;
-  onMouseEnter?: () => void;
-  onMouseLeave?: () => void;
-};
+import { iButtonGroupProps } from '@/types/types/components/components';
 
-const ButtonGroup = ({
-  id,
-  outline,
-  children,
-  className,
-  testId,
-  onClick,
-  onMouseEnter,
-  onMouseLeave,
-  ...rest
-}: iButtonGroupProps) => {
-  return (
-    <FlowbiteButton.Group
-      id={id}
-      outline={outline}
-      className={className}
-      data-testid={testId}
-      onClick={onClick}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-      {...rest}
-    >
-      {children}
-    </FlowbiteButton.Group>
-  );
-};
+const ButtonGroup = forwardRef<any, iButtonGroupProps>(
+  (
+    {
+      id,
+      outline,
+      children,
+      className,
+      testId,
+      onClick,
+      onMouseEnter,
+      onMouseLeave,
+      ...rest
+    }: iButtonGroupProps,
+    ref: React.Ref<any> | null,
+  ) => {
+    return (
+      <FlowbiteButton.Group
+        id={id}
+        outline={outline}
+        className={className}
+        data-testid={testId}
+        onClick={onClick}
+        onMouseEnter={onMouseEnter}
+        onMouseLeave={onMouseLeave}
+        ref={ref}
+        {...rest}
+      >
+        {children}
+      </FlowbiteButton.Group>
+    );
+  },
+);
 
 export default ButtonGroup;
