--- src/types/components/alert.ts
@@ -1,3 +1,0 @@
-export enum AlertType {
-
-}
