--- src/types/components/textbox.ts
@@ -1,6 +1,0 @@
-export enum TextInputType {
-  text = "text",
-  passwords = "passwords",
-  number = "number",
-  email = "email",
-}
